devide modules
