+++ apps/app/src/server/service/yjs/index.ts
@@ -1,0 +1,1 @@
+export * from './yjs';
